backend/requirements.txt
flask
flask_cors
psycopg2-binary
sqlalchemy
geoalchemy2
pandas
xarray
pymongo
dotenv
argopy
google-generativeai
chromadb
scikit-learn
requests
matplotlib
fastapi
uvicorn[standard]
slowapi
pytest